Split sample builder tests by contract
